requirements.txt
flask
flask-cors
supabase
python-dotenv
solana
solders
requests
